New dropdown component and password change success message
